+++ 2011/debuggers_part2_code/x64/bp_use_lib.c
@@ -6,6 +6,7 @@
 #include <stdio.h>
 #include <stdarg.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <string.h>
 #include <signal.h>
 #include <syscall.h>
@@ -20,7 +21,7 @@
 #include "debuglib.h"
 
 
-void run_debugger(pid_t child_pid, long addr)
+void run_debugger(pid_t child_pid, uint64_t addr)
 {
     procmsg("debugger started\n");
 
@@ -74,7 +75,7 @@ int main(int argc, char** argv)
     if (child_pid == 0)
         run_target(argv[1]);
     else if (child_pid > 0) {
-        long addr = strtol(argv[2], NULL, 16);
+        uint64_t addr = (uint64_t) strtol(argv[2], NULL, 16);
         run_debugger(child_pid, addr);
     }
     else {
